Добавлен обработка реакций пользователей на отзывы.

+++ main/src/main/java/ru/yandex/practicum/filmorate/model/UserReactionToReview.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @EqualsAndHashCode
 public class UserReactionToReview {
 
-    private Long reactionId;
+    private Integer reactionId;
     private Integer userId;
     private Integer reviewId;
     private String reaction;
